+++ src/argilla/__init__.py
@@ -44,6 +44,7 @@ if _TYPE_CHECKING:
         delete_records,
         get_workspace,
         init,
+        list_datasets,
         list_workspaces,
         load,
         log,
@@ -119,6 +120,7 @@ _import_structure = {
         "log_async",
         "set_workspace",
         "active_client",
+        "list_datasets",
         "list_workspaces",
     ],
     "client.models": [