+++ src/argilla/client/feedback/training/__init__.py
@@ -13,6 +13,25 @@
 #  limitations under the License.
 
 from argilla.client.feedback.training.base import ArgillaTrainer
-from argilla.client.feedback.training.schemas import TrainingTaskMapping, TrainingTaskMappingForTextClassification
+from argilla.client.feedback.training.schemas import (
+    TrainingTask,
+    TrainingTaskForDPO,
+    TrainingTaskForPPO,
+    TrainingTaskForRM,
+    TrainingTaskForSFT,
+    TrainingTaskForTextClassification,
+    TrainingTaskMapping,  # <- Deprecated
+    TrainingTaskMappingForTextClassification,  # <- Deprecated
+)
 
-__all__ = ["ArgillaTrainer", "TrainingTaskMapping", "TrainingTaskMappingForTextClassification"]
+__all__ = [
+    "ArgillaTrainer",
+    "TrainingTask",
+    "TrainingTaskForTextClassification",
+    "TrainingTaskForSFT",
+    "TrainingTaskForRM",
+    "TrainingTaskForPPO",
+    "TrainingTaskForDPO",
+    "TrainingTaskMapping",
+    "TrainingTaskMappingForTextClassification",
+]