+++ src/argilla/tasks/users/__main__.py
@@ -17,6 +17,8 @@ import typer
 from argilla.tasks.callback import init_callback
 
 from .create import create_user
+from .delete import delete_user
 
 app = typer.Typer(help="Holds CLI commands for user management.", no_args_is_help=True, callback=init_callback)
-app.command(name="create", help="Creates a new user.")(create_user)
+app.command(name="create", help="Creates a new user")(create_user)
+app.command(name="delete", help="Deletes a user")(delete_user)