+++ CHANGELOG.md
@@ -16,6 +16,8 @@ These are the section headers that we use:
 
 ## [Unreleased]()
 
+## [1.15.0](https://github.com/argilla-io/argilla/compare/v1.14.1...v1.15.0)
+
 ### Added
 
 - Added `Enable to update guidelines and dataset settings for Feedback Datasets directly in the UI` ([#3489](https://github.com/argilla-io/argilla/pull/3489))